Refactoring code #63 (#64)
* refactor: extract reusable Badge component for badges

Replace inline `renderBadge` function in `FastestLapItem` with a new `Badge` component and update `CategoryBadge` to use it, reducing code duplication.

* feat(api): add RaceEventPersistence layer for race event queries

Extract database query logic into a dedicated persistence class to improve separation of concerns. Provides methods for finding events by ID, future races, sorted/filtered lists, minimal event data for sitemaps, year statistics, and available years. The new service is registered in the RaceEventsModule.

---------
